Moving streaming configs to Streaming.scala
