+++ core/src/main/scala/org/apache/spark/internal/config/package.scala
@@ -351,51 +351,6 @@ package object config {
     ConfigBuilder("spark.dynamicAllocation.sustainedSchedulerBacklogTimeout")
       .fallbackConf(DYN_ALLOCATION_SCHEDULER_BACKLOG_TIMEOUT)
 
-  private[spark] val STREAMING_DYN_ALLOCATION_ENABLED =
-    ConfigBuilder("spark.streaming.dynamicAllocation.enabled")
-      .booleanConf
-      .createWithDefault(false)
-
-  private[spark] val STREAMING_DYN_ALLOCATION_TESTING =
-    ConfigBuilder("spark.streaming.dynamicAllocation.testing")
-      .booleanConf
-      .createWithDefault(false)
-
-  private[spark] val STREAMING_DYN_ALLOCATION_MIN_EXECUTORS =
-    ConfigBuilder("spark.streaming.dynamicAllocation.minExecutors")
-      .intConf
-      .checkValue(_ > 0, "The min executor number of streaming dynamic " +
-        "allocation must be positive.")
-      .createOptional
-
-  private[spark] val STREAMING_DYN_ALLOCATION_MAX_EXECUTORS =
-    ConfigBuilder("spark.streaming.dynamicAllocation.maxExecutors")
-      .intConf
-      .checkValue(_ > 0, "The max executor number of streaming dynamic " +
-        "allocation must be positive.")
-      .createWithDefault(Int.MaxValue)
-
-  private[spark] val STREAMING_DYN_ALLOCATION_SCALING_INTERVAL =
-    ConfigBuilder("spark.streaming.dynamicAllocation.scalingInterval")
-      .timeConf(TimeUnit.SECONDS)
-      .checkValue(_ > 0, "The scaling interval of streaming dynamic " +
-        "allocation must be positive.")
-      .createWithDefault(60)
-
-  private[spark] val STREAMING_DYN_ALLOCATION_SCALING_UP_RATIO =
-    ConfigBuilder("spark.streaming.dynamicAllocation.scalingUpRatio")
-      .doubleConf
-      .checkValue(_ > 0, "The scaling up ratio of streaming dynamic " +
-        "allocation must be positive.")
-      .createWithDefault(0.9)
-
-  private[spark] val STREAMING_DYN_ALLOCATION_SCALING_DOWN_RATIO =
-    ConfigBuilder("spark.streaming.dynamicAllocation.scalingDownRatio")
-      .doubleConf
-      .checkValue(_ > 0, "The scaling down ratio of streaming dynamic " +
-        "allocation must be positive.")
-      .createWithDefault(0.3)
-
   private[spark] val LOCALITY_WAIT = ConfigBuilder("spark.locality.wait")
     .timeConf(TimeUnit.MILLISECONDS)
     .createWithDefaultString("3s")
